style(client): initial layout design
My initial attempt was to integrate a zero runtime CSS-in-JS lib to
manage app styling, but most of the libs don't work out of the box with CRA.

Since the app is small, let's start with a simple CSS solution.
CRA supports CSS Modules & Scss out of the box, and this combo is
more than enough to manage a small application styling efficiently.